olx-scraper/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
numpy==2.4.2
pandas==3.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
urllib3==2.6.3
